requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
asn1crypto==1.5.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
black==22.10.0
bleach==6.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
coincurve==15.0.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==39.0.0
cytoolz==0.12.0
dataclasses==0.6
dataclassy==0.11.1
Deprecated==1.2.13
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
eciespy==0.3.11
eip712==0.1.0
enforce-typing==1.0.0.post1
eth-abi==2.2.0
eth-account==0.5.9
eth-brownie==1.19.2
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
execnet==1.9.0
frozenlist==1.3.1
gitdb==4.0.10
GitPython==3.1.30
gotrue==0.5.4
h11==0.14.0
hexbytes==0.2.3
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
hypothesis==6.27.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.0
iniconfig==1.1.1
invoke==1.7.3
ipfshttpclient==0.8.0a2
jaraco.classes==3.2.3
json-sempai==0.4.0
jsonschema==3.2.0
keyring==23.13.1
lazy-object-proxy==1.7.1
lru-dict==1.1.8
more-itertools==9.0.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
numpy==1.24.1
oauthlib==3.2.2
ocean-contracts==1.1.8
ocean-lib==2.0.1
packaging==21.3
pandas==1.5.3
parsimonious==0.8.1
pathspec==0.10.1
pkginfo==1.9.6
platformdirs==2.5.2
pluggy==1.0.0
postgrest-py==0.10.4
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.16.0
pydantic==1.9.1
Pygments==2.13.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-dateutil==2.8.2
python-dotenv==0.16.0
python-gitlab==3.12.0
python-semantic-release==7.32.1
pythx==1.6.1
pytz==2022.7.1
PyYAML==5.4.1
readme-renderer==37.3
realtime==0.0.5
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rlp==2.0.1
scipy==1.10.0
semantic-version==2.10.0
semver==2.13.0
shroomdk==1.0.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
storage3==0.3.6
StrEnum==0.4.9
supabase==0.7.1
supafunc==0.2.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tqdm==4.64.1
tweepy==4.12.1
twine==3.8.0
typing_extensions==4.4.0
urllib3==1.26.12
varint==1.0.2
vvm==0.1.0
vyper==0.3.7
wcwidth==0.2.5
web3==5.31.1
webencodings==0.5.1
websockets==10.4
wrapt==1.14.1
yarl==1.8.1
zipp==3.12.0